move windcave into its own file
